+++ CHANGES.md
@@ -4,6 +4,8 @@
 - Don't set `:source-map-path` automatically ([#157](https://github.com/boot-clj/boot-cljs/pull/157), [#136](https://github.com/boot-clj/boot-cljs/issues/136))
 - Use relative path of `.cljs.edn` for asset-path instead of name of the file ([#156](https://github.com/boot-clj/boot-cljs/pull/156), [#152](https://github.com/boot-clj/boot-cljs/issues/152))
 - Use provided `:output-dir` and `:output-to` and set default values based on path of `.cljs.edn`, like previously ([#104](https://github.com/boot-clj/boot-cljs/issues/104))
+- Drop support for `.cljs.edn` `:modules` option (with warning) and instead handle `:output-to` paths in
+`:modules` under `:compiler-options`
 
 **[compare](https://github.com/adzerk-oss/boot-cljs/compare/2.0.0...master)**